:art: :sparkles: :construction: BLoC > MVVM, cambio del nombre de los archivos de los widgets, Filtro tipo Radio Group, Horario de la escuela, generador de horario

+++ app/src/main/java/ziox/ramiro/saes/features/saes/features/schedule/ui/components/ScheduleClassView.kt
@@ -26,7 +26,7 @@ fun ScheduleClassView(
         .padding(top = hourHeight.times((classSchedule.hourRange.start.toDouble() - startHour).toFloat()))
         .height(hourHeight.times(classSchedule.hourRange.duration.toFloat()))
         .fillMaxWidth(),
-    backgroundColor = Color(classSchedule.color),
+    backgroundColor = Color(classSchedule.color.toULong()),
     shape = MaterialTheme.shapes.small,
     elevation = 0.dp
 ) {
